requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==23.2.0
boto3==1.37.29
botocore==1.37.29
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
dill==0.3.9
docker==7.1.0
fastapi==0.115.12
google-pasta==0.2.0
h11==0.14.0
idna==3.10
importlib-metadata==6.11.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.17
numpy==1.26.4
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pathos==0.3.3
platformdirs==4.3.7
pox==0.3.5
ppft==1.7.6.9
protobuf==5.29.4
psutil==7.0.0
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
s3transfer==0.11.4
sagemaker==2.243.0
sagemaker-core==1.0.28
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
starlette==0.46.1
tblib==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
